+++ WeArtTrackingRawDataObject.cs
@@ -3,6 +3,7 @@
 *	https://www.weart.it/
 */
 
+using System;
 using WeArt.Core;
 using WeArt.Messages;
 
@@ -64,6 +65,7 @@ namespace WeArt.Components
             if (type != WeArtClient.MessageType.MessageReceived)
                 return;
 
+            //TD
             if (message is RawDataMessage rawDataMessage)
             {
                 if (rawDataMessage.HandSide != _handSide)
@@ -79,6 +81,31 @@ namespace WeArt.Components
                     default: return;
                 }
                 newSample.Timestamp = rawDataMessage.Timestamp;
+                LastSample = newSample;
+
+                DataReceived?.Invoke(newSample);
+            }
+            //TD PRO
+            else if (message is RawDataTouchDiverPro rawDataTDProMessage)
+            {
+                if (rawDataTDProMessage.HandSide != _handSide)
+                    return;
+
+                TrackingRawData newSample = new TrackingRawData();
+                TrackingRawDataG2 rawDataTDPro;
+                switch (ActuationPoint)
+                {
+                    case ActuationPoint.Thumb: rawDataTDPro = rawDataTDProMessage.Thumb; break;
+                    case ActuationPoint.Index: rawDataTDPro = rawDataTDProMessage.Index; break;
+                    case ActuationPoint.Middle: rawDataTDPro = rawDataTDProMessage.Middle; break;
+                    case ActuationPoint.Annular: rawDataTDPro = rawDataTDProMessage.Annular; break;
+                    case ActuationPoint.Pinky: rawDataTDPro = rawDataTDProMessage.Pinky; break;
+                    case ActuationPoint.Palm: rawDataTDPro = rawDataTDProMessage.Palm; break;
+                    default: return;
+                }
+                newSample.Timestamp = rawDataTDProMessage.Timestamp;
+                newSample.Gyroscope = rawDataTDPro.Gyroscope;
+                newSample.Accelerometer = rawDataTDPro.Accelerometer;
                 LastSample = newSample;
 
                 DataReceived?.Invoke(newSample);